# services/auth-service/internal/handler/rest/user.go
package rest

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/rhythin/bookspot/auth-service/internal/entities/packets"
	"github.com/rhythin/bookspot/services/shared/customlogger"
	"github.com/rhythin/bookspot/services/shared/errhandler"
	"github.com/rhythin/bookspot/services/shared/tracing"
	"go.opentelemetry.io/otel"
)

// Register godoc
// @Summary      Register a new user
// @Description  Create a new user account and returns a temporary token
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        request  body      packets.RegisterRequest  true  "Register Request"
// @Success      201      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Router       /user/register [post]
func (h *handler) Register(w http.ResponseWriter, r *http.Request) (err error) {
	tr := otel.Tracer("auth-handler")
	ctx, span := tr.Start(r.Context(), "Register")
	defer span.End()

	var req packets.RegisterRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		tracing.RecordSpanError(span, err)
		customlogger.S().Warnw("failed to decode request", "error", err)
		return errhandler.NewCustomError(err, http.StatusBadRequest, "Invalid request", false)
	}

	tempToken, err := h.service.CreateUser(ctx, &req)
	if err != nil {
		tracing.RecordSpanError(span, err)
		return err
	}

	messages := map[string]interface{}{
		"message":   "User registered successfully",
		"tempToken": tempToken,
	}

	return sendResponse(w, messages, http.StatusCreated)
}

// Login godoc
// @Summary      User login
// @Description  Authenticate user and returns a temporary token
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        request  body      packets.LoginRequest  true  "Login Request"
// @Success      200      {string}  string
// @Failure      400      {object}  map[string]interface{}
// @Router       /user/login [post]
func (h *handler) Login(w http.ResponseWriter, r *http.Request) (err error) {
	tr := otel.Tracer("auth-handler")
	ctx, span := tr.Start(r.Context(), "Login")
	defer span.End()

	var req packets.LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		tracing.RecordSpanError(span, err)
		customlogger.S().Warnw("failed to decode request", "error", err)
		return errhandler.NewCustomError(err, http.StatusBadRequest, "Invalid request", false)
	}

	tempToken, err := h.service.Login(ctx, &req)
	if err != nil {
		tracing.RecordSpanError(span, err)
		return err
	}

	return sendResponse(w, tempToken, http.StatusOK)
}

// GetUsers godoc
// @Summary      List users
// @Description  Get a paginated list of users
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        limit   query     int     false  "Limit"
// @Param        offset  query     int     false  "Offset"
// @Param        search  query     string  false  "Search term"
// @Success      200     {object}  packets.ListUsersResponse
// @Failure      500     {object}  map[string]interface{}
// @Router       /user [get]
func (h *handler) GetUsers(w http.ResponseWriter, r *http.Request) (err error) {
	tr := otel.Tracer("auth-handler")
	ctx, span := tr.Start(r.Context(), "GetUsers")
	defer span.End()

	limit := r.URL.Query().Get("limit")
	offset := r.URL.Query().Get("offset")
	search := r.URL.Query().Get("search")

	limitInt, err := strconv.Atoi(limit)
	if err != nil {
		customlogger.S().Warnw("failed to convert limit to int", "Error", err)
		customlogger.S().Info("using default limit", "Limit", 10)
		limitInt = 10
	}

	offsetInt, err := strconv.Atoi(offset)
	if err != nil {
		customlogger.S().Warnw("failed to convert offset to int", "Error", err)
		customlogger.S().Info("using default offset", "Offset", 0)
	}

	req := packets.ListUsersRequest{
		Limit:  limitInt,
		Offset: offsetInt,
		Search: search,
	}

	users, err := h.service.GetUsers(ctx, &req)
	if err != nil {
		tracing.RecordSpanError(span, err)
		return err
	}

	return sendResponse(w, users, http.StatusOK)
}

// GetUser godoc
// @Summary      Get user by ID
// @Description  Get detailed information about a specific user
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        userID  path      string  true  "User ID"
// @Success      200     {object}  packets.UserDetails
// @Failure      400     {object}  map[string]interface{}
// @Failure      404     {object}  map[string]interface{}
// @Router       /user/{userID} [get]
func (h *handler) GetUser(w http.ResponseWriter, r *http.Request) (err error) {
	tr := otel.Tracer("auth-handler")
	ctx, span := tr.Start(r.Context(), "GetUser")
	defer span.End()

	userID := chi.URLParam(r, "userID")
	if userID == "" {
		err := errhandler.NewCustomError(errors.New("userID is required"), http.StatusBadRequest, "UserID is required", false)
		tracing.RecordSpanError(span, err)
		return err
	}

	user, err := h.service.GetUser(ctx, userID)
	if err != nil {
		tracing.RecordSpanError(span, err)
		return err
	}
	return sendResponse(w, user, http.StatusOK)
}

// Logout godoc
// @Summary      User logout
// @Description  Log out the current user
// @Tags         user
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Router       /user/logout [post]
func (h *handler) Logout(w http.ResponseWriter, r *http.Request) (err error) {
	tr := otel.Tracer("auth-handler")
	ctx, span := tr.Start(r.Context(), "Logout")
	defer span.End()

	if err := h.service.Logout(ctx); err != nil {
		tracing.RecordSpanError(span, err)
		return err
	}

	messages := map[string]interface{}{
		"message": "User logged out successfully",
	}

	return sendResponse(w, messages, http.StatusOK)
}

// DeleteUser godoc
// @Summary      Delete user
// @Description  Permanently delete a user account
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        userID  path      string  true  "User ID"
// @Success      200     {object}  map[string]interface{}
// @Failure      400     {object}  map[string]interface{}
// @Router       /user/{userID} [delete]
func (h *handler) DeleteUser(w http.ResponseWriter, r *http.Request) (err error) {
	tr := otel.Tracer("auth-handler")
	ctx, span := tr.Start(r.Context(), "DeleteUser")
	defer span.End()

	userID := chi.URLParam(r, "userID")
	if userID == "" {
		err := errhandler.NewCustomError(errors.New("userID is required"), http.StatusBadRequest, "UserID is required", false)
		tracing.RecordSpanError(span, err)
		return err
	}

	if err := h.service.DeleteUser(ctx, userID); err != nil {
		tracing.RecordSpanError(span, err)
		return err
	}

	messages := map[string]interface{}{
		"message": "User deleted successfully",
	}

	return sendResponse(w, messages, http.StatusOK)
}